generate: make the size mandatory
it may still be found by reading the destination file size, but an infinite
stream won't be generated anymore. This is done to simplify the implementation,
and prevent generating an invalid stream that can't be validated.
When stopping the generation because stdout is closed, there is a good chance
that the last chunk won't have its checksum.

Signed-off-by: Gaëtan Lehmann <gaetan.lehmann@vates.tech>